pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arxiv-method-agent"
version = "0.1.0"
description = "LangGraph-based arXiv scouting and benchmark agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.34",
  "docker>=7.1",
  "feedparser>=6.0",
  "langgraph>=1.1,<2",
  "langgraph-checkpoint-sqlite>=2.0",
  "langchain-anthropic>=0.3.0",
  "numpy>=1.26",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "python-dateutil>=2.9",
  "requests>=2.32",
  "tenacity>=9.0",
  "typing-extensions>=4.12",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
